# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Game Top-Up Website API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface StoreSettings {
  storeName: string;
  whatsappNumber: string;
  storeTheme: string;
  packageUnit?: string;
  categories?: string;
  units?: string;
  heroStatus?: string;
  heroTitle1?: string;
  heroTitle2?: string;
  heroSubtitle?: string;
  /** @nullable */
  heroBgImage?: string | null;
}

export interface SettingsInput {
  /** @nullable */
  storeName?: string | null;
  /** @nullable */
  whatsappNumber?: string | null;
  /** @nullable */
  storeTheme?: string | null;
  /** @nullable */
  adminPassword?: string | null;
  /** @nullable */
  packageUnit?: string | null;
  /** @nullable */
  categories?: string | null;
  /** @nullable */
  units?: string | null;
  /** @nullable */
  heroStatus?: string | null;
  /** @nullable */
  heroTitle1?: string | null;
  /** @nullable */
  heroTitle2?: string | null;
  /** @nullable */
  heroSubtitle?: string | null;
  /** @nullable */
  heroBgImage?: string | null;
}

export interface AdminLoginInput {
  password: string;
}

export interface AdminLoginResult {
  success: boolean;
}

export interface Package {
  id: number;
  name: string;
  description: string;
  amount: number;
  price: number;
  currency: string;
  category: string;
  popular: boolean;
  unit?: string;
  createdAt: string;
}

export interface OrderInput {
  packageId: number;
  gameId: string;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  customerNotes?: string | null;
  contactPhone: string;
}

export interface ProofInput {
  proofImageUrl: string;
}

export interface Order {
  id: number;
  packageId: number;
  gameId: string;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  customerNotes?: string | null;
  contactPhone: string;
  status: string;
  /** @nullable */
  proofImageUrl?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface OrderWithPackage {
  id: number;
  packageId: number;
  gameId: string;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  customerNotes?: string | null;
  contactPhone: string;
  status: string;
  /** @nullable */
  proofImageUrl?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
  package: Package;
}

export interface OrderStatusInput {
  status: string;
  /** @nullable */
  notes?: string | null;
}

export interface PackageInput {
  name: string;
  description: string;
  amount: number;
  price: number;
  currency: string;
  category: string;
  popular: boolean;
  unit?: string;
}

export interface PackageUpdate {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  amount?: number | null;
  /** @nullable */
  price?: number | null;
  /** @nullable */
  currency?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  popular?: boolean | null;
  /** @nullable */
  unit?: string | null;
}

export type ListAdminOrdersParams = {
status?: string;
};