# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Game Top-Up Website API
 * OpenAPI spec version: 0.1.0
 */

export interface OrderInput {
  packageId: number;
  gameId: string;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  customerNotes?: string | null;
  contactPhone: string;
}