# lib/api-zod/src/generated/types/orderStatusInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Game Top-Up Website API
 * OpenAPI spec version: 0.1.0
 */

export interface OrderStatusInput {
  status: string;
  /** @nullable */
  notes?: string | null;
}