# lib/api-zod/src/generated/types/orderWithPackage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Game Top-Up Website API
 * OpenAPI spec version: 0.1.0
 */
import type { Package } from './package';

export interface OrderWithPackage {
  id: number;
  packageId: number;
  gameId: string;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  customerNotes?: string | null;
  contactPhone: string;
  status: string;
  /** @nullable */
  proofImageUrl?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
  updatedAt: Date;
  package: Package;
}