# lib/api-zod/src/generated/types/package.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Game Top-Up Website API
 * OpenAPI spec version: 0.1.0
 */

export interface Package {
  id: number;
  name: string;
  description: string;
  amount: number;
  price: number;
  currency: string;
  category: string;
  popular: boolean;
  unit?: string;
  createdAt: Date;
}