# lib/api-zod/src/generated/types/packageUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Game Top-Up Website API
 * OpenAPI spec version: 0.1.0
 */

export interface PackageUpdate {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  amount?: number | null;
  /** @nullable */
  price?: number | null;
  /** @nullable */
  currency?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  popular?: boolean | null;
  /** @nullable */
  unit?: string | null;
}