# lib/api-zod/src/generated/types/settingsInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Game Top-Up Website API
 * OpenAPI spec version: 0.1.0
 */

export interface SettingsInput {
  /** @nullable */
  storeName?: string | null;
  /** @nullable */
  whatsappNumber?: string | null;
  /** @nullable */
  storeTheme?: string | null;
  /** @nullable */
  adminPassword?: string | null;
  /** @nullable */
  packageUnit?: string | null;
  /** @nullable */
  categories?: string | null;
  /** @nullable */
  units?: string | null;
  /** @nullable */
  heroStatus?: string | null;
  /** @nullable */
  heroTitle1?: string | null;
  /** @nullable */
  heroTitle2?: string | null;
  /** @nullable */
  heroSubtitle?: string | null;
  /** @nullable */
  heroBgImage?: string | null;
}