# lib/api-zod/src/generated/types/storeSettings.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Game Top-Up Website API
 * OpenAPI spec version: 0.1.0
 */

export interface StoreSettings {
  storeName: string;
  whatsappNumber: string;
  storeTheme: string;
  packageUnit?: string;
  categories?: string;
  units?: string;
  heroStatus?: string;
  heroTitle1?: string;
  heroTitle2?: string;
  heroSubtitle?: string;
  /** @nullable */
  heroBgImage?: string | null;
}